api/requirements.txt
asgiref==3.5.2
astroid==2.11.7
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.5.1
Django==3.2.15
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.21.3
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
openapi-codec==1.3.2
packaging==21.3
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.3
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
zipp==3.8.1
